+++ spec/Loader/LoaderBehavior.php
@@ -2,10 +2,12 @@
 
 namespace spec\Supervisor\Configuration\Loader;
 
+use Supervisor\Configuration\Loader\LoaderInterface;
+
 trait LoaderBehavior
 {
     function it_is_a_loader()
     {
-        $this->shouldImplement('Supervisor\Configuration\Loader');
+        $this->shouldImplement(LoaderInterface::class);
     }
 }