+++ spec/Section/SectionBehavior.php
@@ -6,6 +6,6 @@ trait SectionBehavior
 {
     function it_is_a_section()
     {
-        $this->shouldImplement('Supervisor\Configuration\Section');
+        $this->shouldImplement('Supervisor\Configuration\Section\SectionInterface');
     }
 }