+++ src/Section/GenericSection.php
@@ -2,14 +2,12 @@
 
 namespace Supervisor\Configuration\Section;
 
-use Supervisor\Configuration\Section;
-
 /**
  * Section name contains identifier.
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class GenericSection implements Section
+final class GenericSection implements SectionInterface
 {
     use SectionData;