+++ src/Writer/AbstractWriter.php
@@ -1,16 +1,13 @@
 <?php
-
 namespace Supervisor\Configuration\Writer;
 
 use Indigo\Ini\Renderer;
+use Supervisor\Configuration\Configuration;
 
-/**
- * Provides an INI Renderer.
- *
- * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
- */
-trait HasIniRenderer
+abstract class AbstractWriter implements WriterInterface
 {
+    abstract public function write(Configuration $configuration);
+
     /**
      * @var Renderer
      */
@@ -19,7 +16,7 @@ trait HasIniRenderer
     /**
      * @return Renderer
      */
-    protected function getRenderer()
+    protected function getRenderer(): Renderer
     {
         if (!isset($this->renderer)) {
             $this->renderer = new Renderer(Renderer::ARRAY_MODE_CONCAT | Renderer::BOOLEAN_MODE_BOOL_STRING);