+++ src/Writer/FlysystemWriter.php
@@ -1,6 +1,5 @@
 <?php
 
-
 namespace Supervisor\Configuration\Writer;
 
 use League\Flysystem\Filesystem;
@@ -12,18 +11,25 @@ use Supervisor\Configuration\Exception\WriterException;
  *
  * @author Márk Sági-Kazár <mark.sagikazar@gmail.com>
  */
-final class IniFileWriter extends AbstractWriter
+final class FlysystemWriter extends AbstractWriter
 {
+    /**
+     * @var Filesystem
+     */
+    private $filesystem;
+
     /**
      * @var string
      */
     private $file;
 
     /**
-     * @param string $file
+     * @param Filesystem $filesystem
+     * @param string     $file
      */
-    public function __construct(string $file)
+    public function __construct(Filesystem $filesystem, $file)
     {
+        $this->filesystem = $filesystem;
         $this->file = $file;
     }
 
@@ -34,7 +40,7 @@ final class IniFileWriter extends AbstractWriter
     {
         $ini = $this->getRenderer()->render($configuration->toArray());
 
-        if (false === $result = file_put_contents($this->file, $ini)) {
+        if (false === $result = $this->filesystem->put($this->file, $ini)) {
             throw new WriterException(sprintf('Cannot write configuration into file %s', $this->file));
         }
     }